BI-3925 refactoring authentication and session middleware and tests
