+++ src/middleware/authentication.middleware.ts
@@ -1,0 +1,11 @@
+import { authMiddleware } from "web-security-node";
+import * as pageURLs from "../model/page.urls";
+
+const authMiddlewareConfig = {
+  accountWebUrl: "",
+  returnUrl: `${pageURLs.STRIKE_OFF_OBJECTIONS}${pageURLs.COMPANY_NUMBER}`,
+};
+
+const signInMiddleware = authMiddleware(authMiddlewareConfig);
+
+export default signInMiddleware;
