+++ test/routes/routes.spec.unit.ts
@@ -1,6 +1,17 @@
+jest.mock("../../src/middleware/authentication.middleware");
+jest.mock("../../src/middleware/session.middleware");
+
+import { NextFunction, Request, Response } from "express";
 import request from "supertest";
 import app from "../../src/app";
+import signInMiddleware from "../../src/middleware/authentication.middleware";
+import sessionMiddleware from "../../src/middleware/session.middleware";
 import { COOKIE_NAME } from "../../src/properties";
+
+const mockSignInMiddleware = signInMiddleware as jest.Mock;
+mockSignInMiddleware.mockImplementation((req: Request, res: Response, next: NextFunction) => { next(); });
+const mockSessionMiddleware = sessionMiddleware as jest.Mock;
+mockSessionMiddleware.mockImplementation((req: Request, res: Response, next: NextFunction) => { next(); });
 
 describe("Basic URL Tests", () => {
